+++ src/data/__init__.py
@@ -1,5 +1,10 @@
 """Local data-interface helpers."""
 
+from data.bluechip_cohort import (
+    BENCHMARK_SYMBOL,
+    BLUECHIP_50_COHORT,
+    get_bluechip_50_symbols,
+)
 from data.csv_loader import (
     CSVValidationSummary,
     ValidatedCSVFrame,
@@ -24,9 +29,17 @@ from data.local_csv_inventory import (
     SUPPORTED_LOCAL_CSV_SCHEMAS,
     validate_local_csv_inventory,
 )
+from data.parquet_loader import (
+    DataIntegrityError,
+    load_eod_cohort_panels,
+    load_eod_parquet,
+)
 
 __all__ = [
+    "BENCHMARK_SYMBOL",
+    "BLUECHIP_50_COHORT",
     "CSVValidationSummary",
+    "DataIntegrityError",
     "LocalCSVInventoryIssue",
     "LocalCSVInventoryReview",
     "LocalCSVInventorySummary",
@@ -37,9 +50,12 @@ __all__ = [
     "ValidatedCSVPanel",
     "ValidatedCSVSeries",
     "generate_diagnostic_cohort_prices",
+    "get_bluechip_50_symbols",
     "load_benchmark_price_csv",
     "load_diagnostic_cohort",
     "load_diagnostic_cohort_manifest",
+    "load_eod_cohort_panels",
+    "load_eod_parquet",
     "load_long_price_csv",
     "load_ohlcv_csv",
     "load_wide_price_csv",